+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
   - Drop dependency on `promise-polyfill`
   - Drop dependency on @braintree/class-list
+  - Update braintree-web to 3.94.0
 
 ## 1.36.1
   - Update braintree-web to 3.92.1